Move relevant logic to hook
